+++ internal/xerrors/errorf_test.go
@@ -15,27 +15,27 @@ func TestErrorf(t *testing.T) {
 		s   string
 	}{
 		{
-			err: MultiErrorf("simple"),
+			err: Errorf("simple"),
 			s:   "simple",
 		},
 		{
-			err: MultiErrorf("format int(%d) and string(%q)", 100500, "azaza"),
+			err: Errorf("format int(%d) and string(%q)", 100500, "azaza"),
 			s:   "format int(100500) and string(\"azaza\")",
 		},
 		{
-			err: MultiErrorf("format int(%d) and string(%q) (error '%w')", 100500, "azaza", context.Canceled),
+			err: Errorf("format int(%d) and string(%q) (error '%w')", 100500, "azaza", context.Canceled),
 			iss: []error{context.Canceled},
 			ass: nil,
 			s:   "format int(100500) and string(\"azaza\") (error 'context canceled')",
 		},
 		{
-			err: MultiErrorf("format int(%d), error '%w' and string(%q)", 100500, context.Canceled, "azaza"),
+			err: Errorf("format int(%d), error '%w' and string(%q)", 100500, context.Canceled, "azaza"),
 			iss: []error{context.Canceled},
 			ass: nil,
 			s:   "format int(100500), error 'context canceled' and string(\"azaza\")",
 		},
 		{
-			err: MultiErrorf("format int(%d), error1 '%w', error2 '%w' and string(%q)",
+			err: Errorf("format int(%d), error1 '%w', error2 '%w' and string(%q)",
 				100500, context.Canceled, context.DeadlineExceeded, "azaza",
 			),
 			iss: []error{context.Canceled, context.DeadlineExceeded},
@@ -43,7 +43,7 @@ func TestErrorf(t *testing.T) {
 			s:   "format int(100500), error1 'context canceled', error2 'context deadline exceeded' and string(\"azaza\")",
 		},
 		{
-			err: MultiErrorf("format int(%d), error1 '%w', error2 '%w' and string(%q)",
+			err: Errorf("format int(%d), error1 '%w', error2 '%w' and string(%q)",
 				100500, Wrap(context.Canceled), context.DeadlineExceeded, "azaza",
 			),
 			iss: []error{context.Canceled, context.DeadlineExceeded},