+++ tests/monitoring_test.go
@@ -18,8 +18,6 @@ import (
 )
 
 func TestMonitoring(t *testing.T) {
-	t.Skip("wait for newest cr.yandex/yc/yandex-docker-local-ydb:latest was published")
-
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
 	defer cancel()