+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 * Added `internal/xerrors.Errorf` error for wrap multiple errors and check them with `errors.Is` of `errors.As`
 * Fixed corner cases of `internal/wait.Wait`
+* Added check of port in connection string and erro throw
+* Fixed bug with initialization of connection pool before apply static credentials
+* Refactored of applying grpc dial options with defaults
 * Added `trace.Driver.{OnBalancerDialEntrypoint,OnBalancerClusterDiscoveryAttempt}` trace events
 
 ## v3.42.8