+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed bug with initialization of connection pool before apply static credentials
 * Refactored of applying grpc dial options with defaults
 * Added `trace.Driver.{OnBalancerDialEntrypoint,OnBalancerClusterDiscoveryAttempt}` trace events
+* Fixed compilation of package `internal/xresolver` with `google.golang.org/grpc@v1.53`
 
 ## v3.42.8
 * Fixed `internal/scheme/helpers/IsDirectoryExists(..)` recursive bug